set/get rundata from Bmob and Pullrefresh

+++ src/com/example/nikirun/HistoryTrackData.java
@@ -166,7 +166,7 @@ public class HistoryTrackData {
         return distance;
     }
 
-    public void setDistance(double distance) {
+    public void setDistance(int distance) {
         this.distance = distance;
     }
